refactor(musicxml): decompose parser and converter into sub-modules
Decompose two monolithic files (~1400 and ~2300 lines) into focused
sub-modules following single-responsibility principle:

Parser (parser.rs -> parser/):
- mod.rs: Document-level orchestration
- attributes.rs: Key, clef, time signature parsing
- measure.rs: Measure element routing
- note.rs: Note, pitch, articulation parsing
- structure.rs: Barline, repeat, direction parsing

Converter (converter.rs -> converter/):
- mod.rs: Entry point, TimingContext, convert/convert_part
- notes.rs: Note collection, conversion, rest distribution
- staff.rs: Single/multi-staff construction
- structure.rs: Pickup detection, repeats, voltas, octave shifts
- ties.rs: Tie and slur chain resolution
- voice.rs: VoiceDistributor for voice assignment

All 57 tests pass (8 new unit tests added). Zero clippy warnings.
No public API changes - all sub-module functions use pub(super).
Each module has module-level doc comments describing its responsibility.

Closes #057
