+++ specs/057-refactor-musicxml-importer/checklists/requirements.md
@@ -1,0 +1,38 @@
+# Specification Quality Checklist: Refactor MusicXML Importer
+
+**Purpose**: Validate specification completeness and quality before proceeding to planning
+**Created**: 2025-03-25
+**Feature**: [spec.md](../spec.md)
+
+## Content Quality
+
+- [x] No implementation details (languages, frameworks, APIs)
+- [x] Focused on user value and business needs
+- [x] Written for non-technical stakeholders
+- [x] All mandatory sections completed
+
+## Requirement Completeness
+
+- [x] No [NEEDS CLARIFICATION] markers remain
+- [x] Requirements are testable and unambiguous
+- [x] Success criteria are measurable
+- [x] Success criteria are technology-agnostic (no implementation details)
+- [x] All acceptance scenarios are defined
+- [x] Edge cases are identified
+- [x] Scope is clearly bounded
+- [x] Dependencies and assumptions identified
+
+## Feature Readiness
+
+- [x] All functional requirements have clear acceptance criteria
+- [x] User scenarios cover primary flows
+- [x] Feature meets measurable outcomes defined in Success Criteria
+- [x] No implementation details leak into specification
+
+## Notes
+
+- All items pass validation.
+- The spec references file sizes (line counts) as evidence of the problem, which is appropriate context for a refactoring spec — these describe the current state, not the implementation approach.
+- FR-007 mentions `NoteData` by name as a domain concept (intermediate type), not as an implementation directive. This is acceptable since it describes *what* to reorganize, not *how*.
+- SC-001's 400-line threshold is a measurable, technology-agnostic metric (lines of source code is a universal measure).
+- The Assumptions section explicitly notes the 400-line target is a guideline, maintaining flexibility.
